requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio>=4.5
black==24.3.0
certifi==2024.2.2
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
distlib==0.3.8
exceptiongroup>=1.2.2
fake-useragent==1.5.1
Faker==24.4.0
filelock==3.13.3
flake8==7.0.0
ffmpeg-python==0.2.0
h11==0.14.0
httpcore==1.0.5
httpx>=0.28.1
identify==2.5.35
idna==3.6
isort==5.13.2
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
parsel==1.9.0
pathspec==0.12.1
platformdirs==4.2.0
pre-commit==3.7.0
pycodestyle==2.11.1
pydantic>=2.11.7
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv>=1.1.0
PyYAML==6.0.1
rich>=13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
tomli==2.0.1
typer>=0.15.2
virtualenv==20.25.1
w3lib==2.1.2
fastmcp==2.10.2